% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y2_artisan_functions copy.R
\name{add1s}
\alias{add1s}
\title{Add 1 Blank PowerPoint slide}
\usage{
add1s(slide_name = "Findings / 1 chart", master_name = "Office Theme")
}
\arguments{
\item{slide_name}{DEFAULT: "Findings / 1 chart"; The name of the type of the PP slide you want added to the PP}

\item{master_name}{DEFAULT: "Office Theme"; The name of the PP master layout that the slide_name comes from}
}
\description{
This function adds a new PowerPoint slide along with a title and commentary box to the current pp object in R
}
\examples{
doc <- add1s()
}
\keyword{powerpoint}
\keyword{slide}
