% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add1t_y2.R
\name{add1t_y2}
\alias{add1t_y2}
\title{Add PowerPoint slide & 1 table}
\usage{
add1t_y2(name, add_slide = T, text_boxes = T,
  slide_name = "Findings / 1 chart", master_name = "Office Theme",
  left_start = 0.5, top_start = 2, height = 5, width = 12)
}
\arguments{
\item{name}{The name of the dataframe to be added to a new PowerPoint slide.}

\item{add_slide}{DEFAULT = T; Automatically adds a blank slide for the chart to be added onto. If F, no new slide is added}

\item{text_boxes}{DEFAULT = T; Automatically adds title and commentary boxes to the blank slide. Set to F for no text boxes}

\item{slide_name}{The name of the type of the PP slide you want added to the PP. DEFAULT: "Findings / 1 chart"}

\item{master_name}{The name of the PP master layout that the slide_name comes from. DEFAULT: "Office Theme"}
}
\description{
This function adds a new PowerPoint slide and fits 1 table onto it. It automatically fits the location to the center of the slide. Beware that if the object passed to the function is not a pre-formated flextable, it automatically assigns a background color to the table and sets the fontsize to 18.
}
\examples{
# First create a table that you can add into a powerpoint object
my_table <- mtcars \%>\%
  y2clerk::freqs(carb) \%>\%
  orderlabel::order_label(inherent_order_label = T)

# Then before adding additional slides, charts, or tables onto a powerpoint, you must first read a powerpoint into R
doc <- read_pptx('~/Dropbox (Y2 Analytics)/Y2 Analytics Team Folder/Resources/Qualtrics Template New.pptx')

# Now start adding in your tables
doc <- add1t_y2(my_table)

print(doc, '~/Desktop/test.pptx')
}
\keyword{chart}
