% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y2_artisan_functions copy.R
\name{add3c}
\alias{add3c}
\title{Add 3 charts to PowerPoint slide}
\usage{
add3c(name, position)
}
\arguments{
\item{name}{The name of the ms_chart object to be added to a new PowerPoint slide.}

\item{position}{Position options: "left"; "center"; "right"; "bottomright"; "topright". The chart layout can either be left-center-right OR left-topright-bottomright.}
}
\description{
This function adds 3 charts to a PowerPoint slide. The charts are automatically added to the last slide of the PP object in R.
}
\examples{
doc <- add_slide(doc, layout = "Findings / 1 chart", master = "Office Theme")
doc <- add3c(chart_name, 'left')
doc <- add3c(chart_name, 'middle')
doc <- add3c(chart_name, 'right')
OR
doc <- add_slide(doc, layout = "Findings / 1 chart", master = "Office Theme")
doc <- add3c(chart_name, 'topright')
doc <- add3c(chart_name, 'bottomright')
doc <- add3c(chart_name, 'left')
}
\keyword{chart}
