% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y2_artisan_functions copy.R
\name{add4c}
\alias{add4c}
\title{Add 4 charts to PowerPoint slide}
\usage{
add4c(name, position)
}
\arguments{
\item{name}{The name of the ms_chart object to be added to a new PowerPoint slide.}

\item{position}{Position options: "topright"; "bottomright"; "topleft"; "bottomleft"; "left"; "centerleft"; "centerright"; "right". The chart layout can either be left-centerleft-centerright-right OR topleft-bottomleft-topright-bottomright.}
}
\description{
This function adds 4 charts to a PowerPoint slide. The charts are automatically added to the last slide of the PP object in R.
}
\examples{
doc <- add_slide(doc, layout = "Findings / 1 chart", master = "Office Theme")
doc <- add4c(chart_name, 'topright')
doc <- add4c(chart_name, 'bottomright')
doc <- add4c(chart_name, 'topleft')
doc <- add4c(chart_name, 'bottomleft')
OR
doc <- add_slide(doc, layout = "Findings / 1 chart", master = "Office Theme")
doc <- add4c(chart_name, 'left')
doc <- add4c(chart_name, 'centerleft')
doc <- add4c(chart_name, 'centerright')
doc <- add4c(chart_name, 'right')
}
\keyword{chart}
