% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y2_artisan_functions copy.R
\name{add5c}
\alias{add5c}
\title{Add 5 charts to PowerPoint slide}
\usage{
add5c(name, position, label_first_only = F)
}
\arguments{
\item{name}{The name of the ms_chart object to be added to a new PowerPoint slide.}

\item{position}{Position options: "left"; "centerleft"; "center"; centerright"; "right". The chart layout is always 5 tall charts}

\item{label_first_only}{DEFAULT = F; Set to T if only the first chart has axis labels. Changing this setting to T in this case will slightly adjut positioning for equally sized graphs}
}
\description{
This function adds 5 vertical charts to a PowerPoint slide. The charts are automatically added to the last slide of the PP object in R.
}
\examples{
doc <- officer::add_slide(doc, layout = "Findings / 1 chart", master = "Office Theme")
doc <- add5c(chart_name, 'left')
doc <- add5c(chart_name, 'centerleft')
doc <- add5c(chart_name, 'center')
doc <- add5c(chart_name, 'centerright')
doc <- add5c(chart_name, 'right')
OR
doc <- officer::add_slide(doc, layout = "Findings / 1 chart", master = "Office Theme")
doc <- add5c(chart_name, 'left', label_first_only = T)
doc <- add5c(chart_name, 'centerleft', label_first_only = T)
doc <- add5c(chart_name, 'center', T)
doc <- add5c(chart_name, 'centerright', T)
doc <- add5c(chart_name, 'right', T)
}
\keyword{chart}
