% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add5c_y2.R
\name{add5c_y2}
\alias{add5c_y2}
\title{Add 5 charts to PowerPoint slide}
\usage{
add5c_y2(name, position, label_first_only = F)
}
\arguments{
\item{name}{The name of the chart/table/other object to be added to a new PowerPoint slide.}

\item{position}{Position options: "left"; "centerleft"; "center"; centerright"; "right". The chart layout is always 5 tall charts}

\item{label_first_only}{DEFAULT = F; Set to T if only the first chart has axis labels. Changing this setting to T in this case will slightly adjut positioning for equally sized graphs}
}
\description{
This function adds 5 vertical charts/tables/other objects to a PowerPoint slide. The charts are automatically added to the last slide of the PP object in R. For this function to work, you need a powerpoint object saved into R called "doc"
}
\examples{
# First create a chart that you can add into a powerpoint object
frequencies <- mtcars \%>\%
  y2clerk::freqs(carb) \%>\%
  orderlabel::order_label(inherent_order_label = T)
color_settings <- list('blue')
text_settings<- list('result' = fp_text(font.size = 20))
chart_name <- y2artisan::ms_single_y2()

# Then before adding additional slides, charts, or tables onto a powerpoint, you must first read a powerpoint into R
doc <- read_pptx('~/Dropbox (Y2 Analytics)/Y2 Analytics Team Folder/Resources/Qualtrics Template New.pptx')

# Now start adding in your charts
doc <- add1s_y2()
doc <- add5c_y2(chart_name, 'left')
doc <- add5c_y2(chart_name, 'centerleft')
doc <- add5c_y2(chart_name, 'center')
doc <- add5c_y2(chart_name, 'centerright')
doc <- add5c_y2(chart_name, 'right')

print(doc, '~/Desktop/test.pptx')
}
\keyword{chart}
