% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y2_artisan_functions copy.R
\name{add6c}
\alias{add6c}
\title{Add 6 charts to PowerPoint slide}
\usage{
add6c(name, position, label_first_only = F)
}
\arguments{
\item{name}{The name of the ms_chart object to be added to a new PowerPoint slide.}

\item{position}{Position options: "topright"; "bottomright"; "topleft"; "bottomleft"; "topcenter"; "bottomcenter". The chart layout can either be left-centerleft-centerright-right OR topleft-bottomleft-topright-bottomright.}

\item{label_first_only}{DEFAULT = F; Set to T for only the first chart to have axis labels. Changing this setting to T in this case will slightly adjut positioning for equally sized graphs}
}
\description{
This function adds 6 charts to a PowerPoint slide. The charts are automatically added to the last slide of the PP object in R.
}
\examples{
doc <- officer::add_slide(doc, layout = "Findings / 1 chart", master = "Office Theme")
doc <- add6c(chart_name, 'topleft')
doc <- add6c(chart_name, 'bottomleft')
doc <- add6c(chart_name, 'topcenter')
doc <- add6c(chart_name, 'bottomcenter')
doc <- add6c(chart_name, 'topright')
doc <- add6c(chart_name, 'bottomright')
}
\keyword{chart}
