% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_colors_facebook_y2.R
\name{add_colors_facebook_y2}
\alias{add_colors_facebook_y2}
\title{Add base Facebook colors}
\usage{
add_colors_facebook_y2(palette = 1, show_colors = TRUE)
}
\arguments{
\item{palette}{DEFAULT = 1. Defaults to the newest facebook color palette (1).}

\item{show_colors}{DEFAULT = TRUE Show the color palettes loaded into your environment}
}
\description{
Primary/Secondary Facebook colors: BLUE_MARKETING, BLUE_FIG, BLUE_HIGHLIGHT, TEAL, LIME, LEMON, ORANGE
}
\details{
Adds to your environment pre-set hex codes for the Facebook blues along with
4 secondary colors used in Facebook themed projects. There are an additional
11 tertiary colors. The 4 blues have 7 shades. Also adds BLACK, WHITE, and
GRAY, while optionally displaying the colors/hex codes for you to see.
}
\examples{
add_colors_facebook_y2()
add_colors_facebook_y2(show_colors = TRUE)
}
\keyword{color}
\keyword{facebook}
\keyword{graph}
