% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_colors_internal_y2.R
\name{add_colors_internal_y2}
\alias{add_colors_internal_y2}
\title{Add base Y2 colors}
\usage{
add_colors_internal_y2(palette = 1, compliments = FALSE, show_colors = TRUE)
}
\arguments{
\item{palette}{DEFAULT = 1. Defaults to the newest Y2 color palette (1). Switch between the latest color palette and older color palettes, with 1 being the oldest. Currently only 1 available.}

\item{compliments}{DEFAULT = FALSE. Add complimentary colors chosen by the Y2 team (Red, Pink, Yellow, Green, Turquoise)}

\item{show_colors}{DEFAULT = TRUE Show the color palettes loaded into your environment}
}
\description{
3 Y2 colors: BLUE, GRAY and ORANGE
}
\details{
Adds to your environment pre-set hex codes for the 3 base colors used
in Y2 themed projects. Each of the 3 colors has 5 shades.
Also adds black and white.
}
\examples{
add_colors_internal_y2()
add_colors_internal_y2(TRUE)
}
\keyword{color}
\keyword{graph}
\keyword{y2}
