% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_colors_microsoft_y2.R
\name{add_colors_microsoft_y2}
\alias{add_colors_microsoft_y2}
\title{Add base Microsoft colors}
\usage{
add_colors_microsoft_y2(palette = c("RE&S")[1], show_colors = TRUE)
}
\arguments{
\item{palette}{DEFAULT = 1. Defaults to the newest microsoft color palette. Switch between the latest color palette and older color palettes.}

\item{show_colors}{DEFAULT = TRUE Show the color palettes loaded into your environment}
}
\description{
Microsoft colors: ORANGE, YELLOW, GREEN, TEAL, BLUE, PURPLE, GRAY
}
\details{
Adds to your environment pre-set hex codes for the 7 base colors used
in Microsoft themed projects. Each of the 7 colors has 3 shades except Gray, which has 6 shades.
Also adds black and white.
}
\examples{
add_colors_microsoft_y2()
add_colors_microsoft_y2(show_colors = TRUE)
}
\keyword{color}
\keyword{graph}
\keyword{microsoft}
