% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_section_header.R
\name{add_section_header}
\alias{add_section_header}
\title{Add Blank PowerPoint Section Header}
\usage{
add_section_header(title = "xxx", slide_name = "Section Header",
  master_name = "Office Theme")
}
\arguments{
\item{title}{DEFAULT: "xxx"; Add a title in quotes here to have your Section Headers put in for you by R}

\item{slide_name}{DEFAULT: "Section Header"; The name of the type of the PP slide you want added to the PP}

\item{master_name}{DEFAULT: "Office Theme"; The name of the PP master layout that the slide_name comes from}
}
\description{
This function adds a new PowerPoint section header along with a title box
}
\examples{
doc <- add_section_header()
OR
add_section_header()
}
\keyword{powerpoint}
\keyword{slide}
