% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_var_label.R
\name{add_var_label}
\alias{add_var_label}
\title{Create custom labels for variables}
\usage{
add_var_label(dataset, label_list)
}
\arguments{
\item{dataset}{A dataframe}

\item{label_list}{A named character vector with variables and their desired labels}
}
\value{
A dataframe
}
\description{
Use add_var_label() to create custom labels for variables in a tidy and pipe friendly manner. Rather than being forced to make permanent changes to your dataset, this will work within a pipe for one time changes or permanent changes if the output is saved to an object. Can assign one or multiple labels using a named vector.
}
\examples{
mtcars \%>\%
  add_var_label(
    c('mpg' = 'Miles per gallon')
  )

mtcars \%>\%
  add_var_label(
    c(
      'mpg' = 'Miles per gallon',
      'hp' = 'Horse Power'
    )
  )
}
\keyword{haven}
\keyword{labelled}
\keyword{labels}
\keyword{var_label}
