% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y2_artisan_functions copy.R
\name{bar_grouped}
\alias{bar_grouped}
\title{Create a groupedbar mschart object}
\usage{
bar_grouped(data = frequencies, x_var = "label", y_var = "result",
  group_var = "group_var", label_text = text_settings_grouped,
  label_color = color_settings_grouped, label_show_values = T,
  label_position = "outEnd", label_num_fmt = "0\%",
  axis_num_fmt = "0\%\%", axis_y_label = "", axis_x_label = "",
  axis_x_display = T, axis_y_min = 0, axis_y_max = NULL,
  axis_y_display = T, axis_text_size = 14, axis_title_size = 18,
  title_label = "", title_size = 18, rotate = 0,
  grouping = "standard", gap_width = 150, overlapping = -50,
  direction = "vertical", legend_text_size = 16, legend_pos = "t")
}
\arguments{
\item{data}{DEFAULT = frequencies; The name of the data frame that the mscharts pulls from.}

\item{x_var}{DEFAULT = 'label'; When using the freqs function, will typically be label (is by default).}

\item{y_var}{DEFAULT = 'result'; When using the freqs function, will typically be result (is by default).}

\item{group_var}{DEFAULT = 'group_var'; All levels of the group_var must be present or the chart may break. To do this, save the variable as_factor() before running freqs. Also remember that label_text and label_color must exactly match all the levels of the group_var or the function will break.}

\item{label_text}{DEFAULT = 'text_settings_grouped'; A list of text settings for the percent labels. This affects font size and color. Specified outside of the function. If a list of one, no need to specify values. Otherwise, they must exactly match the group_var levels. Example: text_settings_grouped <- list('Name of Group 1' = fp_text(font.size = 16, color = lime),'Name of Group 2' = fp_text(font.size = 16, color = brightblue))}

\item{label_color}{DEFAULT = 'color_settings_grouped'; A list of color settings for the bars. This affects font size and color. Specified outside of the function. If a list of one, no need to specify values. Otherwise, they must exactly match the group_var levels. Example: color_settings_grouped <- list('Name of Group 1' = lime,'Name of Group 2' = brightblue)}

\item{label_show_values}{DEFAULT = T; TRUE or FALSE. Show percent labels for each value.}

\item{label_position}{DEFAULT = 'outEnd'; Specifies the position of the data label. It should be one of 'b', 'ctr', 'inBase', 'inEnd', 'l', 'outEnd', 'r', 't'. When grouping is 'clustered', it should be one of 'ctr','inBase','inEnd','outEnd'. When grouping is 'stacked', it should be one of 'ctr','inBase','inEnd'. When grouping is 'standard', it should be one of 'b','ctr','l','r','t'.}

\item{label_num_fmt}{DEFAULT = '0\%'; Number formatting specifies number format properties which indicate how to format and render the numeric values. It can be "General", "0.00", "#,##0", "#,##0.00", "mm-dd-yy", "m/d/yy h:mm", etc.}

\item{axis_num_fmt}{DEFAULT = '0\%\%'; Unlike label_num_fmt, the default for percentages is "0\%\%".}

\item{axis_y_label}{DEFAULT = ''; Title for the y_axis}

\item{axis_x_label}{DEFAULT = ''; Title for the x_axis}

\item{axis_x_display}{DEFAULT = T}

\item{axis_y_min}{DEFAULT = 0 to show full data without skewing perspective, but can be adjusted.}

\item{axis_y_max}{DEFAULT = NULL}

\item{axis_y_display}{DEFAULT = T}

\item{axis_text_size}{DEFAULT = 14; Font size for variable levels and percentages.}

\item{axis_title_size}{DEFAULT = 18; Font size for axis_x_label and axis_y_label.}

\item{title_label}{DEFAULT = ''; Add the question wording from the survey in "" as the title of the chart.}

\item{title_size}{DEFAULT = 18}

\item{rotate}{DEFAULT = 0; Rotation of x_axis text. Set to -45 for diagonal, giving more space for text.}

\item{grouping}{DEFAULT = 'standard'; grouping for a barchart, a linechart or an area chart. must be one of "percentStacked", "clustered", "standard" or "stacked".}

\item{gap_width}{DEFAULT = 150, meaning the size of the space between bars is 150\% the size of the bar itself}

\item{overlapping}{DEFAULT = -50 This leaves 50\% extra space between variable levels. Set to 100 when coloring bars different colors.}

\item{direction}{DEFAULT = 'vertical'; Two options: "vertical" (default) OR "horizontal"}

\item{legend_text_size}{DEFAULT = 16}

\item{legend_pos}{DEFAULT = 't' for top; Other legend positions are 'b', 'tr', 'l', 'r', and 'n' for none.}
}
\description{
This function creates a mschart object automatically formatted for a single variable (including multiple select). It requires two lists called "text_settings" and "color_settings" by default that specify the colors desired for the chart.
}
\keyword{chart}
\keyword{grouped}
