% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y2_artisan_functions copy.R
\name{bar_single}
\alias{bar_single}
\title{Create an ugrouped mschart object}
\usage{
bar_single(data = frequencies, x_var = "label", y_var = "result",
  group_var = NULL, label_text = text_settings,
  label_color = color_settings, label_show_values = T,
  label_position = "outEnd", label_num_fmt = "0\%",
  axis_num_fmt = "0\%\%", axis_x_label = "", axis_y_label = "",
  axis_y_min = 0, axis_y_max = NULL, axis_text_size = 14,
  axis_title_size = 18, title_label = "", title_size = 18,
  rotate = 0, grouping = "standard", overlapping = -50,
  direction = "vertical", legend_text_size = 16, legend_pos = "n")
}
\arguments{
\item{data}{The name of the data frame that the mscharts pulls from.}

\item{x_var}{When using the freqs function, will typically be label (is by default).}

\item{y_var}{When using the freqs function, will typically be result (is by default).}

\item{group_var}{Is null by default. If you want the bars to be different colors, set group_var to the same variable as x_var. Then set overlap to 100.}

\item{label_text}{A list of text settings for the percent labels. This affects font size and color. Specified outside of the function. If a list of one, no need to specify values. Otherwise, they must exactly match the group_var levels. Example: text_settings <- list(fp_text(font.size = 10.5, color = bluepurple))}

\item{label_color}{A list of color settings for the bars. This affects font size and color. Specified outside of the function. If a list of one, no need to specify values. Otherwise, they must exactly match the group_var levels. Example: color_settings <- list(bluepurple)}

\item{label_show_values}{TRUE or FALSE. Show percent labels for each value.}

\item{label_position}{Specifies the position of the data label. It should be one of 'b', 'ctr', 'inBase', 'inEnd', 'l', 'outEnd', 'r', 't'. When grouping is 'clustered', it should be one of 'ctr','inBase','inEnd','outEnd'. When grouping is 'stacked', it should be one of 'ctr','inBase','inEnd'. When grouping is 'standard', it should be one of 'b','ctr','l','r','t'.}

\item{label_num_fmt}{Number formatting specifies number format properties which indicate how to format and render the numeric values. It can be "General", "0.00", "#,##0", "#,##0.00", "mm-dd-yy", "m/d/yy h:mm", etc.}

\item{axis_num_fmt}{Unlike label_num_fmt, the default for percentages is "0\%\%".}

\item{axis_x_label}{Title for the x_axis. Blank by default.}

\item{axis_y_label}{Title for the y_axis. Blank by default.}

\item{axis_y_min}{Set to 0 to show full data without skewing perspective, but can be adjusted.}

\item{axis_y_max}{No maximum specified by default.}

\item{axis_text_size}{Font size for variable levels and percentages.}

\item{axis_title_size}{Font size for axis_x_label and axis_y_label.}

\item{title_label}{Add the question wording from the survey in "" as the title of the chart.}

\item{rotate}{Rotation of x_axis text. Set to -45 for diagonal, giving more space for text.}

\item{grouping}{grouping for a barchart, a linechart or an area chart. must be one of "percentStacked", "clustered", "standard" or "stacked".}

\item{overlapping}{Set to -50 by default. This leaves 50\% extra space between variable levels. Set to 100 when coloring bars different colors.}

\item{direction}{Two options: "vertical" (default) OR "horizontal"}

\item{legend_pos}{Set to 'n' for none. Other legend positions are 'b', 'tr', 'l', 'r', 't'.}
}
\description{
This function creates a mschart object automatically formatted for a single variable (including multiple select). It requires two lists called "text_settings" and "color_settings" by default that specify the colors desired for the chart.
}
\examples{
my_ms_chart <- bar_single()
OR
my_ms_chart <- bar_single(
  rotate = -45,
  title_label  = 'Add the question wording from the survey here?',
  group_var = 'label',
  overlap = 100
)
}
\keyword{chart}
