% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_donut_y2.R
\name{gg_donut_y2}
\alias{gg_donut_y2}
\title{Create a donut chart ggplot object}
\usage{
gg_donut_y2(
  data = frequencies,
  color_var = label,
  y_var = result,
  colors = "#ffffff",
  fills,
  font_family = "flama",
  hole_size = 2,
  label_size = 5,
  legend_pos = "none",
  legend_rev = FALSE,
  legend_text_size = 8,
  title_label = "",
  title_size = 14
)
}
\arguments{
\item{data}{DEFAULT = frequencies; The name of the data frame that ggplot pulls from}

\item{color_var}{DEFAULT = label; Note: the color variable CANNOT be numeric}

\item{y_var}{DEFAULT = result; When using the freqs function, will typically be result (is by default)}

\item{colors}{DEFAULT is white ('#ffffff') for the text of all percent labels; You may also 1) Specify 1 color, and this color will be applied to all color_var levels or 2) Specify a vector of colors for each individual level of the color_var}

\item{fills}{NO DEFAULT; requires a vector of colors for all levels of the x_var}

\item{font_family}{DEFAULT = 'flama'; all fonts used need to be previously loaded in using the font_add() and showtext_auto() functions}

\item{hole_size}{DEFAULT = 3. The larger the size, the larger the donut hole}

\item{label_size}{DEFAULT = 10. Adjusts the size of the percent labels over each bar}

\item{legend_pos}{DEFAULT = 'none'}

\item{legend_rev}{DEFAULT = FALSE}

\item{legend_text_size}{DEFAULT = 8}

\item{title_label}{DEFAULT = ''; Add your title in "" as the title of the chart}

\item{title_size}{DEFAULT = 18}
}
\description{
This function creates a ggplot2 object automatically formatted for a donut chart
}
\examples{
frequencies <- iris \%>\% y2clerk::freqs(Species)

chart <- gg_donut_y2(
  fills = c('red', 'blue', 'pink'),
  font_family = 'sans'
)
}
\keyword{bar}
\keyword{chart}
\keyword{ggplot}
\keyword{single}
