% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_single.R
\name{gg_single}
\alias{gg_single}
\title{Create an ugrouped ggplot object}
\usage{
gg_single(data = frequencies, x_var = label, y_var = result,
  label_var = percent_label, color_var = label, axis_text_size = 12,
  axis_title_size = 14, bar_width = 0.75, direction = "vertical",
  fills = rep("#474E7E", count(data)), label_length = 45,
  label_size = 10, legend_pos = "none", legend_rev = FALSE,
  legend_text_size = 6, legend_title = element_blank(), nudge = 0,
  text_family = "flama", title_label = "", title_size = 14,
  x_label = "", y_label = "", y_min = 0, y_max = 0)
}
\arguments{
\item{data}{DEFAULT = frequencies;The name of the data frame that ggplot pulls from.}

\item{x_var}{DEFAULT = label; When using the freqs function, will typically be label (is by default).}

\item{y_var}{DEFAULT = result; When using the freqs function, will typically be result (is by default).}

\item{label_var}{DEFAULT = percent_label; When using the order_label function, this variable will be created for you.}

\item{color_var}{DEFAULT = label; Although the color_var is set to label, the default for the "fills" argument sets all bars to be the same color.}

\item{axis_text_size}{DEFAULT = 12; Font size for variable levels and percentages.}

\item{axis_title_size}{DEFAULT = 14; Font size for x_label and y_label.}

\item{bar_width}{DEFAULT = .75, with a bar_width of 1 meaning each bars touches the ones next to it}

\item{direction}{DEFAULT = 'vertical'; Two options: "vertical" (default) OR "horizontal"}

\item{fills}{DEFAULT = rep('#474E7E', count(data)); the hexcode is the bluepurple color from the Qualtrics logo. This short function will give all the bars the same color. If colors are changed here in conjunction with the color_var, you can more easily manipulate colors based on another variable.}

\item{label_length}{DEFAULT = 45 for horizontal charts and 15 for vertical charts}

\item{label_size}{DEFAULT = 10}

\item{legend_pos}{DEFAULT = 'none'}

\item{legend_text_size}{DEFAULT = 6}

\item{legend_title}{DEFAULT = element_blank()}

\item{nudge}{DEFAULT = 0; however, nudge automatically adjusts based on the max value of 'result', in most cases fitting the chart perfectly}

\item{text_family}{DEFAULT = 'flama'}

\item{title_label}{DEFAULT = ''; Add the question wording from the survey in "" as the title of the chart.}

\item{title_size}{DEFAULT = 18}

\item{y_label}{DEFAULT = ''; Title for the y_axis}

\item{y_min}{DEFAULT = 0 to show full data without skewing perspective, but can be adjusted.}

\item{y_max}{DEFAULT = 0; however, the y_max automatically adjusts based on the max value of 'result', in most cases fitting the chart perfectly}
}
\description{
This function creates a mschart object automatically formatted for a single variable (including multiple select). It requires two lists called "text_settings" and "color_settings" by default that specify the colors desired for the chart.
}
\examples{
my_ms_chart <- bar_single()
OR
my_ms_chart <- bar_single(
  rotate = -45,
  title_label  = 'Add the question wording from the survey here?',
  group_var = 'label',
  overlap = 100
)
}
\keyword{bar}
\keyword{chart}
\keyword{ggplot}
\keyword{single}
