% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_stacked_y2.R
\name{gg_stacked_y2}
\alias{gg_stacked_y2}
\title{Create a grouped ggplot object}
\usage{
gg_stacked_y2(
  data = frequencies,
  x_var = group_var,
  y_var = result,
  label_var = percent_label,
  color_var = label,
  axis_display = T,
  axis_text_size = 12,
  axis_title_size = 14,
  bar_width = 0.75,
  direction = "horizontal",
  erase_labels = 0.01,
  fills,
  colors = "0",
  label_length = 45,
  label_size = 8,
  legend_nrow = NULL,
  legend_pos = "top",
  legend_rev = FALSE,
  legend_text_size = 6,
  legend_title_size = 8,
  legend_title = "",
  text_family = "flama",
  title_label = "",
  title_size = 14,
  x_label = "",
  y_label = "",
  y_min = 0
)
}
\arguments{
\item{data}{DEFAULT = frequencies; The name of the data frame that ggplot pulls from.}

\item{x_var}{DEFAULT = group_var; When doing a single stacked bar, set x_var to variable or any variable that has a only one level}

\item{y_var}{DEFAULT = result; When using the freqs function, will typically be result (is by default).}

\item{label_var}{DEFAULT = percent_label; When using the order_label function, this variable will be created for you.}

\item{color_var}{DEFAULT = label. Note: the color variable CANNOT be numeric.}

\item{axis_display}{DEFAULT = TRUE; for a single stacked bar, set to FALSE to remove axis labels}

\item{axis_text_size}{DEFAULT = 12; Font size for variable levels and axis percentages.}

\item{axis_title_size}{DEFAULT = 14; Font size for x_label and y_label.}

\item{bar_width}{DEFAULT = .75, with a bar_width of 1 meaning each bars touches the ones next to it}

\item{erase_labels}{DEFAULT = .01; all percent labels less than or equal to erase_labels will be erased to avoid clutter and overlapping labels. This argument pulls from the value in the result column of the dataframe being used}

\item{fills}{NO DEFAULT; requires a vector of colors for all levels of the color_var}

\item{colors}{DEFAULT is white ('#ffffff') for the text of all percent labels; make custom by inputting a vector of colors for all levels of the color_var}

\item{label_length}{DEFAULT = 45 for horizontal charts and 15 for vertical charts. This determines how many characters an x-axis label can be before R inserts a line break.}

\item{label_size}{DEFAULT = 8. Adjusts the size of the percent labels within each bar.}

\item{legend_nrow}{DEFAULT = NULL; Change to a numeric to specify the number of rows for the legend}

\item{legend_pos}{DEFAULT = 'top'}

\item{legend_rev}{DEFAULT = FALSE}

\item{legend_text_size}{DEFAULT = 6}

\item{legend_title_size}{DEFAULT = 8}

\item{legend_title}{DEFAULT = '', If you put in a title, the legend will default to 'top' unless otherwise specified}

\item{text_family}{DEFAULT = 'flama'; all fonts used need to be loaded in using the theme() and showtext_auto() functions}

\item{title_label}{DEFAULT = ''; Add your title in "" as the title of the chart.}

\item{title_size}{DEFAULT = 18}

\item{y_label}{DEFAULT = ''; Title for the y_axis}

\item{y_min}{DEFAULT = 0. Change to a negative value such as -.2 to add space for extra text/graphic between the axis text and the stacked bars}
}
\description{
This function creates a ggplot2 object automatically formatted for a stacked variable with inherent ordering.
}
\examples{
frequencies <- tibble(
  label = rep(c('Promoter', 'Passive', 'Detractor'), 3),
  result = c(.33, .33, .34, .20, .30, .50, .25, .50, .25),
  value = rep(c(1, 2, 3), 3),
  group_var = c(rep('Group A', 3), rep('Group B', 3), rep('Group C', 3))
) \%>\% order_label(
  group_var = group_var,
  stacked = 'gg'
)

chart <- gg_stacked_y2(
  fills = c('red', 'yellow', 'green')
)


# For a single stacked bar:
frequencies <- tibble(
 label = c('Promoter', 'Passive', 'Detractor'),
 result = c(.33, .33, .34),
 value = c(1, 2, 3),
 variable = rep('QNPS', 3)
) \%>\% order_label(stacked = 'gg')

chart <- gg_stacked_y2(
   x_var = variable,
   axis_display = F,
   fills = c('red', 'yellow', 'green')
)
}
\keyword{bar}
\keyword{chart}
\keyword{ggplot}
\keyword{stacked}
