% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggchart_save_y2.R
\name{ggchart_save_y2}
\alias{ggchart_save_y2}
\title{Quickly save out a chart from ggplot}
\usage{
ggchart_save_y2(
  chart,
  chartname,
  chart_path = CHART_PATH,
  width = 11,
  height = 5.5
)
}
\arguments{
\item{chart}{The ggplot object you want to use.}

\item{chartname}{The name you want to give your file, ex: "brands by age". Do not specify the whole file path. However, you do need to have an object that is the path to your chart folder saved in R as CHART_PATH}

\item{chart_path}{DEFAULT = CHART_PATH. If you set a CHART_PATH at the beginning of your code, you won't have to worry about this again}

\item{width}{DEFAULT = 11}

\item{height}{DEFAULT = 5.5}
}
\description{
Saves a ggplot chart. All you need to add is the file name and a CHART_PATH.
}
\examples{
# To save your chart, you will need CHART_PATH, a path to your folder destination
\dontrun{
CHART_PATH <- '~/Desktop/'

frequencies <- mtcars \%>\%
  y2clerk::freqs(carb) \%>\%
  orderlabel::order_label(inherent_order_label = TRUE)
chart <- gg_single_y2(font_family = "sans")

ggchart_save_y2('test')
}
}
\keyword{chart}
\keyword{ggplot}
\keyword{save}
