% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_chart.R
\name{line_chart}
\alias{line_chart}
\title{Create an ugrouped mschart object}
\usage{
line_chart(data = frequencies, x_var = "year", y_var = "result",
  group_var = "label", label_text = text_settings,
  label_color = color_settings, label_show_values = T,
  label_position = "t", label_num_fmt = "0\%",
  axis_num_fmt = "0\%\%", axis_x_display = T, axis_x_label = "",
  axis_y_label = "", axis_y_min = 0, axis_y_max = NULL,
  axis_y_display = T, axis_text_size = 14, axis_title_size = 18,
  title_label = "", title_size = 18, rotate = 0,
  legend_text_size = 16, legend_pos = "n")
}
\arguments{
\item{data}{DEFAULT = frequencies;The name of the data frame that the mscharts pulls from.}

\item{x_var}{DEFAULT = 'year'; The name of your time variable}

\item{y_var}{DEFAULT = 'result'; When using the freqs function, will typically be result (is by default).}

\item{group_var}{DEFAULT = 'label'; Each line will be a distinct value of this variable}

\item{label_text}{DEFAULT = 'text_settings'; A list of text settings for the percent labels. This affects font size and color. Specified outside of the function. If a list of one, no need to specify values. Otherwise, they must exactly match the group_var levels. Example: text_settings <- list(fp_text(font.size = 10.5, color = bluepurple))}

\item{label_color}{DEFAULT = 'color_settings'; A list of color settings for the bars. This affects font size and color. Specified outside of the function. If a list of one, no need to specify values. Otherwise, they must exactly match the group_var levels. Example: color_settings <- list('District 1' = lime, 'District 2' = bluepurple)}

\item{label_show_values}{DEFAULT = T; TRUE or FALSE. Show percent labels for each value.}

\item{label_position}{DEFAULT = 't'; Specifies the position of the data label. It should be one of 'b', 'ctr', 'inBase', 'inEnd', 'l', 'outEnd', 'r', 't'. When grouping is 'clustered', it should be one of 'ctr','inBase','inEnd','outEnd'. When grouping is 'stacked', it should be one of 'ctr','inBase','inEnd'. When grouping is 'standard', it should be one of 'b','ctr','l','r','t'.}

\item{label_num_fmt}{DEFAULT = '0\%'; Number formatting specifies number format properties which indicate how to format and render the numeric values. It can be "General", "0.00", "#,##0", "#,##0.00", "mm-dd-yy", "m/d/yy h:mm", etc.}

\item{axis_num_fmt}{DEFAULT = '0\%\%'; Unlike label_num_fmt, the default for percentages is "0\%\%".}

\item{axis_x_label}{DEFAULT = ''; Title for the x_axis}

\item{axis_y_label}{DEFAULT = ''; Title for the y_axis}

\item{axis_y_min}{DEFAULT = 0 to show full data without skewing perspective, but can be adjusted.}

\item{axis_y_max}{DEFAULT = NULL}

\item{axis_y_display}{DEFAULT = T}

\item{axis_text_size}{DEFAULT = 14; Font size for variable levels and percentages.}

\item{axis_title_size}{DEFAULT = 18; Font size for axis_x_label and axis_y_label.}

\item{title_label}{DEFAULT = ''; Add the question wording from the survey in "" as the title of the chart.}

\item{legend_text_size}{DEFAULT = 14}

\item{legend_pos}{DEFAULT = 'n' for none; Other legend positions are 'b', 'tr', 'l', 'r', 't'.}
}
\description{
This function creates a mschart object automatically formatted for a single variable (including multiple select). It requires two lists called "text_settings" and "color_settings" by default that specify the colors desired for the chart.
}
\keyword{chart}
