% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_fonts_y2.R
\name{list_fonts_y2}
\alias{list_fonts_y2}
\title{Lists the fonts loaded in by add_fonts_y2}
\usage{
list_fonts_y2()
}
\value{
A dataframe
}
\description{
Use list_fonts_y2() to see the available fonts loaded in by add_fonts_y2().
}
\examples{
list_fonts_y2()
}
\keyword{font}
\keyword{showtext}
