% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpg}
\alias{mpg}
\title{Example data from ggplot2 package}
\format{
A data frame with 234 observations and 11 variables
}
\usage{
mpg
}
\description{
This ggplot2 data is great for practice use in visualizations (used in function examples for y2artisan)
}
\keyword{datasets}
