% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_footer_y2.R
\name{ms_footer_y2}
\alias{ms_footer_y2}
\title{Create a string with question wording and n-size}
\usage{
ms_footer_y2(dataset, variable, q_type = c("s", "m"))
}
\arguments{
\item{dataset}{Haven labelled dataset from which to pull the variable}

\item{variable}{Unquoted variable}

\item{q_type}{DEFAULT = 's'; Can be any of c('s', 'm') used to specify question type. s = single select, m = multiple select, etc. Meant to match Y2 question naming conventions.}
}
\description{
Pipe in the dataset (responses), specify the variable, and set the question type. Save out the nicely formatted string to an object to pass to the footer of an ms chart.
}
\examples{
\dontrun{
responses \%>\%
  ms_footer_y2(
   s_income
  )

mtcars \%>\%
  ms_footer_y2(
   m_aware,
   q_type = 'm'
  )
  }
}
\keyword{footer}
\keyword{n}
\keyword{n-size}
\keyword{question}
