% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_grouped_y2.R
\name{ms_grouped_y2}
\alias{ms_grouped_y2}
\title{Create a groupedbar mschart object}
\usage{
ms_grouped_y2(
  data = frequencies,
  x_var = "label",
  y_var = "result",
  group_var = "group_var",
  axis_text_size = 14,
  axis_title_size = 18,
  axis_x_display = T,
  axis_x_label = "",
  axis_x_position = c("nextTo", "high", "low", "none"),
  axis_x_rotate = 0,
  axis_x_rotate_title = 0,
  axis_y_display = T,
  axis_y_label = "",
  axis_y_min = 0,
  axis_y_max = NULL,
  axis_y_rotate = 0,
  axis_y_rotate_title = 360,
  direction = c("vertical", "horizontal"),
  font_family = "Arial",
  gap_width = 75,
  grouping = "standard",
  label_color = color_settings_grouped,
  label_position = c("outEnd", "inEnd", "ctr", "inBase"),
  label_show_values = T,
  label_text = text_settings_grouped,
  overlapping = -25,
  legend_pos = c("t", "n", "b", "tr", "l", "r"),
  legend_text_size = 16,
  num_fmt = c("percent", "general"),
  title_label = "",
  title_size = 18
)
}
\arguments{
\item{data}{DEFAULT = frequencies; The name of the data frame that the mscharts pulls from.}

\item{x_var}{DEFAULT = 'label'; When using the freqs function, will typically be label (is by default).}

\item{y_var}{DEFAULT = 'result'; When using the freqs function, will typically be result (is by default).}

\item{group_var}{DEFAULT = 'group_var'; All levels of the group_var must be present or the chart may break. To do this, save the variable as_factor() before running freqs. Also remember that label_text and label_color must exactly match all the levels of the group_var or the function will break.}

\item{axis_text_size}{DEFAULT = 14; Font size for variable levels and percentages.}

\item{axis_title_size}{DEFAULT = 18; Font size for axis_x_label and axis_y_label.}

\item{axis_x_position}{DEFAULT = 'nextTo'; Other options include "high", "low", "none". Change to "low" if dealing with negative numbers}

\item{axis_y_display}{DEFAULT = T}

\item{axis_y_label}{DEFAULT = ''; Title for the y_axis}

\item{axis_y_min}{DEFAULT = 0 to show full data without skewing perspective, but can be adjusted.}

\item{axis_y_max}{DEFAULT = NULL}

\item{axis_y_rotate}{DEFAULT = 0; Rotation of y_axis text. Set to -45 for diagonal, giving more space for text.}

\item{axis_y_rotate_title}{DEFAULT = 360, default for x_axis is 0}

\item{direction}{DEFAULT = 'vertical'; Two options: "vertical" (default) OR "horizontal"}

\item{font_family}{DEFAULT = 'Arial'. Sets the fonts for axis, legends, and titles. Label font is set within label_color and label_text lists. May specify fonts in quotes, e.g. "Times New Roman"}

\item{gap_width}{DEFAULT = 25, meaning the size of the space between bars is 25\% the size of the bar itself}

\item{grouping}{DEFAULT = 'standard'; grouping for a barchart, a linechart or an area chart. must be one of "percentStacked", "clustered", "standard" or "stacked".}

\item{label_color}{DEFAULT = color_settings_grouped; A list of color settings for the bars. This affects font size and color. Specified outside of the function. If a list of one, no need to specify values. Otherwise, they must exactly match the group_var levels. Example: color_settings_grouped <- list('Name of Group 1' = lime,'Name of Group 2' = brightblue)}

\item{label_position}{DEFAULT = 'outEnd'; Other options include c('outEnd', 'inEnd', 'ctr', 'inBase')}

\item{label_show_values}{DEFAULT = T; TRUE or FALSE. Show percent labels for each value.}

\item{label_text}{DEFAULT = text_settings_grouped; A list of text settings for the percent labels. This affects font size and color. Specified outside of the function. If a list of one, no need to specify values. Otherwise, they must exactly match the group_var levels. Example: text_settings_grouped <- list('Name of Group 1' = fp_text(font.size = 16, color = lime),'Name of Group 2' = fp_text(font.size = 16, color = brightblue))}

\item{overlapping}{DEFAULT = -50 This leaves 50\% extra space between variable levels. Set to 100 when coloring bars different colors.}

\item{legend_pos}{DEFAULT = 't' for top; Other legend positions are 'b', 'tr', 'l', 'r', and 'n' for none.}

\item{legend_text_size}{DEFAULT = 16}

\item{num_fmt}{DEFAULT = 'percent'; Can also be set to 'general' for non-percentages. Changes formatting for both the labels and axis}

\item{title_label}{DEFAULT = ''; Add your title in "" as the title of the chart.}

\item{title_size}{DEFAULT = 18}
}
\description{
This function creates a mschart object automatically formatted for a grouped variable. It requires two lists called "text_settings_grouped" and "color_settings_grouped" by default that specify the colors desired for the chart.
}
\examples{
frequencies <- ToothGrowth \%>\%
  group_by(supp) \%>\%
  freqs(dose) \%>\%
  order_label(group_var = group_var)

color_settings_grouped <- list(
  'OJ' = 'orange',
  'VC' = 'gray'
)
text_settings_grouped <- list(
  'OJ' = fp_text(color = 'orange', font.size = 20),
  'VC' = fp_text(color = 'gray', font.size = 20)
)

chart <- ms_grouped_y2()
print(chart, preview = T)
}
\keyword{chart}
\keyword{grouped}
