% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_line_y2.R
\name{ms_line_y2}
\alias{ms_line_y2}
\title{Create an ugrouped mschart object}
\usage{
ms_line_y2(
  data = frequencies,
  x_var = "label",
  y_var = "result",
  group_var = "group_var",
  axis_text_size = 14,
  axis_title_size = 18,
  axis_x_display = T,
  axis_x_label = "",
  axis_x_rotate = 0,
  axis_x_rotate_title = 0,
  axis_y_display = T,
  axis_y_label = "",
  axis_y_min = 0,
  axis_y_max = NULL,
  axis_y_rotate = 0,
  axis_y_rotate_title = 360,
  font_family = "Arial",
  label_color = color_settings_grouped,
  label_position = "t",
  label_show_values = T,
  label_text = text_settings_grouped,
  legend_pos = c("n", "t", "b", "tr", "l", "r"),
  legend_text_size = 16,
  num_fmt = c("percent", "general"),
  title_label = "",
  title_size = 18
)
}
\arguments{
\item{data}{DEFAULT = frequencies;The name of the data frame that the mscharts pulls from.}

\item{x_var}{DEFAULT = 'label'; The name of your time variable}

\item{y_var}{DEFAULT = 'result'; When using the freqs function, will typically be result (is by default).}

\item{group_var}{DEFAULT = 'group_var'; Each line will be a distinct value of this variable}

\item{axis_text_size}{DEFAULT = 14; Font size for variable levels and percentages.}

\item{axis_title_size}{DEFAULT = 18; Font size for axis_x_label and axis_y_label.}

\item{axis_y_display}{DEFAULT = T}

\item{axis_y_label}{DEFAULT = ''; Title for the y_axis}

\item{axis_y_min}{DEFAULT = 0 to show full data without skewing perspective, but can be adjusted.}

\item{axis_y_max}{DEFAULT = NULL}

\item{axis_y_rotate}{DEFAULT = 0; Rotation of y_axis text. Set to -45 for diagonal, giving more space for text.}

\item{axis_y_rotate_title}{DEFAULT = 360, default for x_axis is 0}

\item{font_family}{DEFAULT = 'Arial'. Sets the fonts for axis, legends, and titles. Label font is set within label_color and label_text lists. May specify fonts in quotes, e.g. "Times New Roman"}

\item{label_color}{DEFAULT = color_settings_grouped; A list of color settings for the bars. This affects font size and color. Specified outside of the function. If a list of one, no need to specify values. Otherwise, they must exactly match the group_var levels. Example: color_settings <- list('District 1' = lime, 'District 2' = bluepurple)}

\item{label_position}{DEFAULT = 't'; Specifies the position of the data label. It should be one of 'b', 'ctr', 'inBase', 'inEnd', 'l', 'outEnd', 'r', 't'. When grouping is 'clustered', it should be one of 'ctr','inBase','inEnd','outEnd'. When grouping is 'stacked', it should be one of 'ctr','inBase','inEnd'. When grouping is 'standard', it should be one of 'b','ctr','l','r','t'.}

\item{label_show_values}{DEFAULT = T; TRUE or FALSE. Show percent labels for each value.}

\item{label_text}{DEFAULT = text_settings_grouped; A list of text settings for the percent labels. This affects font size and color. Specified outside of the function. If a list of one, no need to specify values. Otherwise, they must exactly match the group_var levels. Example: text_settings <- list(fp_text(font.size = 10.5, color = bluepurple))}

\item{legend_pos}{DEFAULT = 'n' for none; Other legend positions are 'b', 'tr', 'l', 'r', 't'.}

\item{legend_text_size}{DEFAULT = 14}

\item{num_fmt}{DEFAULT = 'percent'; Can also be set to 'general' for non-percentages. Changes formatting for both the labels and axis}

\item{title_label}{DEFAULT = ''; Add your title in "" as the title of the chart.}

\item{title_size}{DEFAULT = 18}
}
\description{
This function creates a mschart object automatically formatted for a line chart. It requires two lists called "text_settings_grouped" and "color_settings_grouped" by default that specify the colors desired for the chart.
}
\examples{
frequencies <- tibble(
  label = rep(c('2000', '2001', '2002'), 2),
  group_var = rep(c('Brand 1', 'Brand 2'), 3),
  result = c(.18, .20, .24, .25, .24, .23)
)

color_settings_grouped <- list(
  'Brand 1' = 'green',
  'Brand 2' = 'red'
)
text_settings_grouped <- list(
  'Brand 1' = fp_text(color = 'green'),
  'Brand 2' = fp_text(color = 'red')
)

chart <- ms_line_y2()
print(chart, preview = T)
}
\keyword{chart}
