% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_print_y2.R
\name{ms_print_y2}
\alias{ms_print_y2}
\title{Print an ms_chart object for your viewing pleasure}
\usage{
ms_print_y2(chart_name = chart)
}
\arguments{
\item{chart_name}{DEFAULT = chart; A chart object made using y2artisan or ms_charts}
}
\description{
Provide a chart object created using the y2artisan suite of ms_charts function to see a preview of the chart.
}
\examples{

frequencies <- mtcars \%>\%
  y2clerk::freqs(carb) \%>\%
  orderlabel::order_label(inherent_order_label = TRUE)

color_settings <- list('blue')
text_settings<- list('result' = officer::fp_text(font.size = 20))

chart <- ms_single_y2()
ms_print_y2()
}
\keyword{charts}
\keyword{ms}
\keyword{preview}
\keyword{print}
