% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y2_artisan_functions copy.R
\name{openend}
\alias{openend}
\title{Look at the frequencies of each word in an open end question}
\usage{
openend(variable, dataset = responses)
}
\arguments{
\item{variable}{The name of the openended variable from your dataset you want to look at}

\item{dataset}{DEFAULT = responses}
}
\description{
Breaks down an open ended question on spaces, giving you the frequencies of each word mentioned
}
\examples{
frequencies <- openend(QOPEN_END)
}
\keyword{end}
\keyword{freqs}
\keyword{frequencies}
\keyword{open}
\keyword{openend}
