% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_colors.R
\docType{data}
\name{purple}
\alias{purple}
\title{Colors}
\format{An object of class \code{character} of length 1.}
\usage{
purple
}
\description{
Default Qualtrics Colors
}
\details{
Pre saved hex codes for the 4 base colros typically used in Qualtrics projects
}
\keyword{datasets}
