% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_labels.R
\name{restore_labels}
\alias{restore_labels}
\title{Restore original labels to your dataset}
\usage{
restore_labels(dataset, data_labels)
}
\arguments{
\item{dataset}{A dataframe}

\item{data_labels}{A dataframe with a column "name" containing variable names and a column "labels" containing labels. Designed to work with the output of taking_names()}
}
\value{
A dataframe
}
\description{
Use restore_labels() in conjunction with taking_names() to restore variable labels to a dataset.
}
\examples{
mtcars \%>\%
  restore_labels(
    orderlabel::taking_names(mtcars)
  )
}
\keyword{haven}
\keyword{labelled}
\keyword{labels}
\keyword{taking_names}
\keyword{var_label}
