% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_color_settings.R
\name{set_color_settings}
\alias{set_color_settings}
\title{Setup the color setting for ms charts}
\usage{
set_color_settings(..., dataset = frequencies)
}
\arguments{
\item{...}{Colors to use in your ms_chart}

\item{dataset}{DEFAULT = frequencies; A frequency table named frequencies}
}
\description{
Save a frequencies object using freqs(). Provide colors to the set_color_settings function. The colors will be automatically matched to the "label" column of the frequencies.
}
\examples{

add_colors_internal_y2()

frequencies <-
  mtcars \%>\%
  y2clerk::freqs(
    cyl
  )

set_color_settings(
  BLUE_DARKER,
  BLUE,
  BLUE_LIGHTER
 )
}
\keyword{charts}
\keyword{color}
\keyword{ms}
\keyword{settings}
