% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_color_settings_y2.R
\name{set_color_settings_y2}
\alias{set_color_settings_y2}
\title{Setup the color setting for ms charts}
\usage{
set_color_settings_y2(..., dataset = frequencies, color_column = label)
}
\arguments{
\item{...}{Colors to use in your ms_chart}

\item{dataset}{DEFAULT = frequencies; A frequency table named frequencies}

\item{color_column}{DEFAULT = label; Column from which labels are pulled}
}
\description{
Save a frequencies object using freqs(). Provide colors to the set_color_settings function. The colors will be automatically matched to the "label" column of the frequencies or you can manually choose a different column .
}
\examples{

add_colors_internal_y2()

frequencies <-
  mtcars \%>\%
  y2clerk::freqs(
    cyl
  )

set_color_settings_y2(
  BLUE_DARKER,
  BLUE,
  BLUE_LIGHTER
 )

frequencies <-
  mtcars \%>\%
  dplyr::group_by(
    gear
  ) \%>\%
  y2clerk::freqs(
    cyl
  )

set_color_settings_y2(
  BLUE_DARKER,
  BLUE,
  BLUE_LIGHTER,
  color_column = group_var
 )
}
\keyword{charts}
\keyword{color}
\keyword{ms}
\keyword{settings}
