% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_text_settings.R
\name{set_text_settings_y2}
\alias{set_text_settings_y2}
\title{Setup the text setting for ms charts}
\usage{
set_text_settings_y2(
  dataset = frequencies,
  font_size = 14,
  font_color = "white",
  font_family = "BentonSans"
)
}
\arguments{
\item{dataset}{DEFAULT = frequencies; A frequency table named frequencies}

\item{font_size}{DEFAULT = 14; Font size applied to all text in chart labels}

\item{font_color}{DEFAULT = 'white'; Color applied to all text in chart labels}

\item{font_family}{DEFAULT = 'BentonSans'; Font family applied to all text in chart labels}
}
\description{
Save a frequencies object using freqs(). The text settings will be automatically matched to the "label" column of the frequencies.
}
\examples{

frequencies <-
  mtcars \%>\%
  y2clerk::freqs(
    cyl
  )

set_text_settings_y2()

set_text_settings_y2(
  font_size = 28,
  font_color = 'black'
 )
}
\keyword{charts}
\keyword{ms}
\keyword{settings}
\keyword{text}
