% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud.R
\name{wordcloud}
\alias{wordcloud}
\title{Creates a word cloud from an open end question}
\usage{
wordcloud(variable, colors = "bluepurple", dataset = responses)
}
\arguments{
\item{variable}{The name of the openended variable from your dataset you want to look at}

\item{colors}{DEFAULT = 'bluepurple'; 4 qualtrics colors as pre-made options: "bluepurple", "lime", "teal", "brightblue". May also specify a vector of 3 scaled colors ranging from lightest to darkest}

\item{dataset}{DEFAULT = responses}
}
\description{
Finds the frequencies of each word in an open end question and creates a word cloud based on the frequencies. Words mentioned less are smaller and lighter in color
}
\examples{
chart <- wordcloud(QOPEN_END)
chart <- wordcloud(QOPEN_END, lime)
chart <- wordcloud(QOPEN_END, c('#FAEFF2', '#CC6078', '#6D0018'))
}
\keyword{cloud}
\keyword{end}
\keyword{open}
\keyword{openend}
\keyword{word}
\keyword{wordcloud}
