% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud_y2.R
\name{wordcloud_y2}
\alias{wordcloud_y2}
\title{Creates a word cloud from an open end question}
\usage{
wordcloud_y2(dataset, variable, colors = "bluepurple", top_x = 50,
  min_size = 1, max_size = 12)
}
\arguments{
\item{dataset}{no default. Usually piped in from your main dataset}

\item{variable}{The name of the openended variable from your dataset you want to look at}

\item{colors}{DEFAULT = 'bluepurple'; 4 qualtrics colors as pre-made options: "bluepurple", "lime", "teal", "brightblue". May also specify a vector of 3 scaled colors ranging from lightest to darkest}
}
\description{
Finds the frequencies of each word in an open end question and creates a word cloud based on the frequencies. Words mentioned less are smaller and lighter in color
}
\examples{
chart <- wordcloud_y2(QOPEN_END)
chart <- wordcloud_y2(QOPEN_END, lime)
chart <- wordcloud_y2(QOPEN_END, c('#FAEFF2', '#CC6078', '#6D0018'))
}
\keyword{cloud}
\keyword{end}
\keyword{open}
\keyword{openend}
\keyword{word}
\keyword{wordcloud}
