% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add1s_y2.R
\name{add1s_y2}
\alias{add1s_y2}
\title{Add 1 Blank PowerPoint slide}
\usage{
add1s_y2(text_boxes = F, slide_name = "Findings / 1 chart",
  master_name = "Office Theme")
}
\arguments{
\item{text_boxes}{DEFAULT = T; Automatically adds title, commentary, and footer boxes to the blank slide. Set to F for no text boxes}

\item{slide_name}{DEFAULT: "Findings / 1 chart"; The name of the type of the PP slide you want added to the PP}

\item{master_name}{DEFAULT: "Office Theme"; The name of the PP master layout that the slide_name comes from}
}
\description{
This function adds a new PowerPoint slide along with a title and commentary box to the current pp object in R
}
\examples{
# Before adding additional slides, charts, or tables onto a powerpoint, you must first read a powerpoint into R
doc <- read_pptx('~/Dropbox (Y2 Analytics)/Y2 Analytics Team Folder/Resources/Qualtrics Template New.pptx')

doc <- add1s_y2()

print(doc, '~/Desktop/test.pptx')
}
\keyword{powerpoint}
\keyword{slide}
