% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add2t_y2.R
\name{add2t_y2}
\alias{add2t_y2}
\title{Add 2 tables to PowerPoint slide}
\usage{
add2t_y2(name, position)
}
\arguments{
\item{name}{The name of the ms_chart object to be added to a new PowerPoint slide.}

\item{position}{Position options: "top"; "bottom"; "left"; "right". The chart layout can either be top-bottom OR left-right.}
}
\description{
This function adds 2 tables to a PowerPoint slide. The tables are automatically added to the last slide of the PP object in R. Beware that if the object passed to the function is not a pre-formated flextable, it automatically assigns a background color to the table and sets the fontsize to 18.
}
\examples{
# First create a table that you can add into a powerpoint object
my_table <- mtcars \%>\%
  y2clerk::freqs(carb) \%>\%
  orderlabel::order_label(inherent_order_label = T)

# Then before adding additional slides, charts, or tables onto a powerpoint, you must first read a powerpoint into R
doc <- read_pptx('~/Dropbox (Y2 Analytics)/Y2 Analytics Team Folder/Resources/Qualtrics Template New.pptx')

# Now start adding in your tables
doc <- add1s_y2()
doc <- add2t_y2(my_table, 'left')
doc <- add2t_y2(my_table, 'right')

doc <- add1s_y2()
doc <- add2t_y2(my_table, 'top')
doc <- add2t_y2(my_table, 'bottom')

print(doc, '~/Desktop/test.pptx')
}
\keyword{chart}
