% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y2artisan.R
\name{add3c}
\alias{add3c}
\title{Add 3 charts to PowerPoint slide}
\usage{
add3c(name, position, label_first_only = F)
}
\arguments{
\item{name}{The name of the ms_chart object to be added to a new PowerPoint slide.}

\item{position}{Position options: "left"; "center"; "right"; "bottomright"; "topright". The chart layout can either be left-center-right OR left-topright-bottomright.}

\item{label_first_only}{DEFAULT = F; Set to T if for a series of left-to-right charts where only the first chart has axis labels. Changing this setting to T in this case will slightly adjut positioning for equally sized graphs}
}
\description{
This function adds 3 charts to a PowerPoint slide. The charts are automatically added to the last slide of the PP object in R.
}
\examples{
doc <- officer::add_slide(doc, layout = "Findings / 1 chart", master = "Office Theme")
doc <- add3c(chart_name, 'left')
doc <- add3c(chart_name, 'center')
doc <- add3c(chart_name, 'right')
OR
doc <- officer::add_slide(doc, layout = "Findings / 1 chart", master = "Office Theme")
doc <- add3c(chart_name, 'left', label_first_only = T)
doc <- add3c(chart_name, 'middle', T)
doc <- add3c(chart_name, 'right', T)
OR
doc <- officer::add_slide(doc, layout = "Findings / 1 chart", master = "Office Theme")
doc <- add3c(chart_name, 'topright')
doc <- add3c(chart_name, 'bottomright')
doc <- add3c(chart_name, 'left')
}
\keyword{chart}
