% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_colors_y2.R
\name{add_colors_y2}
\alias{add_colors_y2}
\title{Add base Qulatrics colors}
\usage{
add_colors_y2()
}
\description{
4 Qualtrics colors: purple, green, blue, and teal
}
\details{
Adds to your environment pre-set hex codes for the 4 base colors typically used in Qualtrics projects. Each of the 4 colors also has 5 shades. There are no arguments to this function.
}
\examples{
add_colors_y2()
}
\keyword{Qualtrics}
\keyword{color}
\keyword{graph}
