% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_section_header_y2.R
\name{add_section_header_y2}
\alias{add_section_header_y2}
\title{Add Blank PowerPoint Section Header}
\usage{
add_section_header_y2(
  title = "xxx",
  slide_name = "Section Header",
  master_name = "Office Theme"
)
}
\arguments{
\item{title}{DEFAULT: "xxx"; Add a title in quotes here to have your Section Headers put in for you by R}

\item{slide_name}{DEFAULT: "Section Header"; The name of the type of the PP slide you want added to the PP}

\item{master_name}{DEFAULT: "Office Theme"; The name of the PP master layout that the slide_name comes from}
}
\description{
This function adds a new PowerPoint section header along with a title box
}
\examples{
# Before adding additional slides, charts, or tables onto a powerpoint, you must first read a powerpoint into R
doc <- read_pptx('~/Dropbox (Y2 Analytics)/Y2 Analytics Team Folder/Resources/Qualtrics Template New.pptx')

doc <- add_section_header_y2()

print(doc, '~/Desktop/test.pptx')
}
\keyword{powerpoint}
\keyword{slide}
