% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_stacked_y2.R
\name{ms_stacked_y2}
\alias{ms_stacked_y2}
\title{Create a groupedbar mschart object}
\usage{
ms_stacked_y2(data = frequencies, x_var = "group_var",
  y_var = "result", group_var = "label", axis_text_size = 14,
  axis_title_size = 18, axis_x_text_color = "black",
  axis_x_display = T, axis_x_label = "", axis_x_rotate = 0,
  axis_x_rotate_title = 0, axis_y_display = F, axis_y_label = "",
  axis_y_min = 0, axis_y_max = 1, axis_y_rotate = 0,
  axis_y_rotate_title = 360, direction = "horizontal",
  gap_width = 25, grouping = "percentStacked", font_family = "Arial",
  label_color = color_settings_stacked, label_position = "outEnd",
  label_show_percent = F, label_show_values = T,
  label_text = text_settings_stacked, legend_pos = "t",
  legend_text_size = 10, num_fmt = "percent", overlapping = 100,
  title_label = "", title_size = 18)
}
\arguments{
\item{data}{DEFAULT = frequencies; The name of the data frame that the mscharts pulls from.}

\item{x_var}{DEFAULT = 'group_var'; For a single stacked bar, use x_var = 'variable' or 'stat', really anything that will only have 1 level.}

\item{y_var}{DEFAULT = 'result'; When using the freqs function, will typically be result (is by default).}

\item{group_var}{DEFAULT = 'label'; When using the freqs function, will typically be label (is by default). All levels of the group_var must be present or the chart may break. To do this, save the variable as_factor() before running freqs. Also remember that label_text and label_color must exactly match all the levels of the group_var or the function will break.}

\item{axis_text_size}{DEFAULT = 14; Font size for variable levels and percentages.}

\item{axis_title_size}{DEFAULT = 18; Font size for axis_x_label and axis_y_label.}

\item{axis_x_text_color}{DEFAULT = 'black'; Set to 'transparent' for no text on single bars}

\item{axis_y_display}{DEFAULT = F; set axis_x_display to F when it is a single stacked bar}

\item{axis_y_label}{DEFAULT = ''; Title for the y_axis}

\item{axis_y_min}{DEFAULT = 0 to show full data without skewing perspective, but can be adjusted.}

\item{axis_y_max}{DEFAULT = 1 to allow percent totals to add to 100\%.}

\item{axis_y_rotate}{DEFAULT = 0; Rotation of y_axis text. Set to -45 for diagonal, giving more space for text.}

\item{axis_y_rotate_title}{DEFAULT = 360, default for x_axis is 0}

\item{direction}{DEFAULT = 'horizontal'; Two options: "horizontal" (default) OR "vertical"}

\item{gap_width}{DEFAULT = 25, meaning the size of the space between bars is 25\% the size of the bar itself}

\item{grouping}{DEFAULT = 'percentStacked'; grouping for a barchart, a linechart or an area chart. must be one of "percentStacked", "clustered", "standard" or "stacked".}

\item{font_family}{DEFAULT = 'Arial'. Sets the fonts for axis, legends, and titles. Label font is set within label_color and label_text lists. May specify fonts in quotes, e.g. "Times New Roman"}

\item{label_color}{DEFAULT = 'color_settings_stacked'; A list of color settings for the levels within each stacked bar. This affects font size and color. Specified outside of the function. If a list of one, no need to specify values. Otherwise, they must exactly match the group_var levels. Example: color_settings_grouped <- list('Name of Group 1' = lime,'Name of Group 2' = brightblue)}

\item{label_position}{DEFAULT = 'outEnd'; Specifies the position of the data label. It should be one of 'b', 'ctr', 'inBase', 'inEnd', 'l', 'outEnd', 'r', 't'. When grouping is 'clustered', it should be one of 'ctr','inBase','inEnd','outEnd'. When grouping is 'stacked', it should be one of 'ctr','inBase','inEnd'. When grouping is 'standard', it should be one of 'b','ctr','l','r','t'.}

\item{label_show_percent}{DEFAULT = F}

\item{label_show_values}{DEFAULT = T; TRUE or FALSE. Show percent labels for each value.}

\item{label_text}{DEFAULT = 'text_settings_stacked'; A list of text settings for the percent labels. This affects font size and color. Specified outside of the function. If a list of one, no need to specify values. Otherwise, they must exactly match the group_var levels. Example: text_settings_grouped <- list('Name of Group 1' = fp_text(font.size = 16, color = lime),'Name of Group 2' = fp_text(font.size = 16, color = brightblue))}

\item{legend_pos}{DEFAULT = 't' for top; Other legend positions are 'b', 'tr', 'l', 'r', and 'n' for none.}

\item{legend_text_size}{DEFAULT = 10}

\item{num_fmt}{DEFAULT = 'percent'; Can also be set to 'general' for non-percentages. Changes formatting for both the labels and axis}

\item{overlapping}{DEFAULT = 100}

\item{title_label}{DEFAULT = ''; Add your title in "" as the title of the chart.}

\item{title_size}{DEFAULT = 18}
}
\description{
This function creates a mschart object automatically formatted for a single variable (including multiple select). It requires two lists called "text_settings" and "color_settings" by default that specify the colors desired for the chart.
}
\examples{
my_ms_chart <- ms_stacked_y2()

OR for a single stacked bar:
my_ms_chart <- ms_stacked_y2(
  title_label  = 'Add chart title here',
  axis_x_text_color = 'transparent',
  x_var = 'variable'
)
}
\keyword{chart}
\keyword{stacked}
