% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud_y2.R
\name{wordcloud_y2}
\alias{wordcloud_y2}
\title{Creates a word cloud from an open end question}
\usage{
wordcloud_y2(
  dataset,
  variable,
  colors = "bluepurple",
  max_size = 12,
  min_size = 1,
  font_family = "flama",
  top_x = 50
)
}
\arguments{
\item{dataset}{no default. Usually piped in from your main dataset}

\item{variable}{The name of the openended variable from your dataset you want to look at}

\item{colors}{DEFAULT = 'bluepurple'; 4 qualtrics colors as pre-made options: "bluepurple", "lime", "teal", "brightblue". May also specify a vector of 3 scaled colors ranging from lightest to darkest}

\item{max_size}{DEFAULT = 12; the largest text size for the word with the highest frequency}

\item{min_size}{DEFAULT = 1; the smallest text size for the word with the lowest frequency}

\item{font_family}{DEFAULT = 'flama'; all fonts used need to be previously loaded in using the font_add() and showtext_auto() functions}

\item{top_x}{DEFAULT = 50; Shows the top X most commonly mentioned words you want to see from the open-end}
}
\description{
Finds the frequencies of each word in an open end question and creates a word cloud based on the frequencies. Words mentioned less are smaller and lighter in color
}
\examples{
responses <- tibble::tibble(
  var1 = c(
    'I like to talk about dogs',
    'Dogs are cool but cats are aight too',
    'I prefer dogs over cats',
    "My dog's collars are always too tight",
    'One last sentence about dogs',
    'Cats collars are typically cooler than dogs'
  )
)

RED_DARK <- '#C61616'
RED_MID <- '#E38B8B'
RED_LIGHT <- '#F9E6E6'

# If you want to use a specific color, you must specify three scalar shades of that color
responses \%>\% wordcloud_y2(var1, font_family = "Arial")
responses \%>\% wordcloud_y2(var1, c(RED_LIGHT, RED_MID, RED_DARK), font_family = "Arial",)
}
\keyword{cloud}
\keyword{end}
\keyword{open}
\keyword{openend}
\keyword{word}
\keyword{wordcloud}
